[feat] ErrorResponse와 ErrorMessage 구현 #10
