--- week2/src/main/java/server/sopt/week2/dto/ErrorMessage.java
@@ -1,17 +1,0 @@
-package server.sopt.week2.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@AllArgsConstructor
-@Getter
-public enum ErrorMessage {
-
-    MEMBER_NOT_FOUND_BY_ID_EXCEPTION(HttpStatus.NOT_FOUND.value(), "ID에 해당하는 사람이 없습니다"),
-    BLOG_NOT_FOUND_BY_ID_EXCEPTION(HttpStatus.NOT_FOUND.value(), "ID에 해당하는 블로그가 없습니다"),
-
-    ;
-    private final int status;
-    private final String message;
-}
