+++ week2/src/main/java/server/sopt/week2/exception/NotFoundException.java
@@ -1,6 +1,6 @@
 package server.sopt.week2.exception;
 
-import server.sopt.week2.dto.ErrorMessage;
+import server.sopt.week2.error.ErrorMessage;
 
 public class NotFoundException extends BusinessException {
     public NotFoundException(ErrorMessage errorMessage) {
